requirements.txt
cryptography==41.0.7
Pillow==10.0.0 